refactor: split CastVoteCard into card and neuron-select
